item 수정

+++ itemservice/src/main/java/com/ecommerce/itemservice/grpcservice/ItemServiceGrpcServer.java
@@ -30,7 +30,7 @@ public class ItemServiceGrpcServer extends ItemInfoServiceGrpc.ItemInfoServiceIm
                 if (itemDTO != null) {
                         try {
                                 reply = ItemReply.newBuilder()
-                                                .setItemId((int) itemDTO.getItemValues().getItemId())
+                                                .setItemId(itemDTO.getItemValues().getItemId())
                                                 .setCategoryId(itemDTO.getItemValues().getCategory())
                                                 .setItemCount(itemDTO.getSalesValues().getItemCount())
                                                 .setItemPrice(itemDTO.getSalesValues().getItemPrice())
